arm64: dts: ti: k3-j721e-sk: Add IMX219 RPi camera overlay
Two RPi camera headers present on J721E SK and are connected to each
port of two CSI2RX instances 0 and 1. Sensors are connected to main_i2c3
interface via an I2C switch. Both MIPI and RPi connectors are connected
to the CSI2 interface. A mux select should be set to high to select the
RPi connector. Add overlay for enabling the RPi camera connectors and
instantiate two IMX219 RPi cameras on the same.

Signed-off-by: Sinthu Raja <sinthu.raja@ti.com>
Signed-off-by: Vaishnav Achath <vaishnav.a@ti.com>
Signed-off-by: Vignesh Raghavendra <vigneshr@ti.com>