arm64: dts: ti: k3-j721s2-common-proc-board: Enable main_i2c5 and disable csi ports
Enable main_i2c5 instance which is the camera control I2C interface
for the CSI2 instances on J721S2 and also enable the TCA6408 GPIO
expander on main_i2c5. CSI2RX instances which are routed to the Samtec
MIPI connector on the common processor board thus enable them as well.

Signed-off-by: Vaishnav Achath <vaishnav.a@ti.com>
Signed-off-by: Vignesh Raghavendra <vigneshr@ti.com>